+++ previous-versions/assess/src/codatassess/models/operations/generate_excel_report.py
@@ -3,9 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import excelreporttype as shared_excelreporttype
-from ..shared import excelstatus as shared_excelstatus
+from ...models.shared import excelreporttype as shared_excelreporttype
+from ...models.shared import excelstatus as shared_excelstatus
 from typing import Optional
 
 
@@ -23,13 +22,11 @@ class GenerateExcelReportRequest:
 class GenerateExcelReportResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     excel_status: Optional[shared_excelstatus.ExcelStatus] = dataclasses.field(default=None)
     r"""OK"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""