+++ previous-versions/assess/src/codatassess/models/operations/get_data_integrity_summaries.py
@@ -3,9 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import dataintegritydatatype as shared_dataintegritydatatype
-from ..shared import errormessage as shared_errormessage
-from ..shared import summaries as shared_summaries
+from ...models.shared import dataintegritydatatype as shared_dataintegritydatatype
+from ...models.shared import summaries as shared_summaries
 from typing import Optional
 
 
@@ -25,12 +24,10 @@ class GetDataIntegritySummariesRequest:
 class GetDataIntegritySummariesResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
     summaries: Optional[shared_summaries.Summaries] = dataclasses.field(default=None)
     r"""OK"""